backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
httpx>=0.26.0
python-dotenv>=1.0.0
pydantic>=2.5.0
aiohttp>=3.9.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
